Add a test-suite which performs integration/stress tests which load and handle all stored test files in various ways.
It works by using handlers for each type of file which perform various operations on the files, e.g. loading,
iterating content, modify, ... This will trigger changes which break working with the available test-files and
thus provides another layer of regression testing which hopefully prevents some failures from making it into
releases.

It is runnable via a new ant-target 'test-integration' and also added to the jenkins-target.

git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1647885 13f79535-47bb-0310-9956-ffa450edef68